auditd_file_events: Add mmap, read and write rows (see details).
Read and write operations are now included in the table; note that
the mmap() syscall right now is mapped to either a read or write
operation depending on the memory protection flags.

There's a lot of spam right now, since there is no filtering. The
next step would be to deduplicate these rows so that we only get
state change notifications:

1. Show the file as soon as it is read or written
2. Show the file again if the next access type is different than
   the previous one.
3. Silence the file until it is closed (?).